+++ examples/chat-bot/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## [1.15.14](https://github.com/AIGNE-io/aigne-framework/compare/example-chat-bot-v1.15.13...example-chat-bot-v1.15.14) (2025-07-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/cli bumped to 1.24.0
+
+## [1.15.13](https://github.com/AIGNE-io/aigne-framework/compare/example-chat-bot-v1.15.12...example-chat-bot-v1.15.13) (2025-07-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/cli bumped to 1.23.1
+
+## [1.15.12](https://github.com/AIGNE-io/aigne-framework/compare/example-chat-bot-v1.15.11...example-chat-bot-v1.15.12) (2025-07-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/cli bumped to 1.23.0
+
 ## [1.15.11](https://github.com/AIGNE-io/aigne-framework/compare/example-chat-bot-v1.15.10...example-chat-bot-v1.15.11) (2025-07-14)