+++ packages/transport/CHANGELOG.md
@@ -1,5 +1,53 @@
 # Changelog
 
+## [0.10.2](https://github.com/AIGNE-io/aigne-framework/compare/transport-v0.10.1...transport-v0.10.2) (2025-07-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/openai bumped to 0.10.2
+  * devDependencies
+    * @aigne/agent-library bumped to 1.21.2
+    * @aigne/core bumped to 1.36.0
+    * @aigne/test-utils bumped to 0.5.10
+    * @aigne/default-memory bumped to 1.0.2
+
+## [0.10.1](https://github.com/AIGNE-io/aigne-framework/compare/transport-v0.10.0...transport-v0.10.1) (2025-07-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/openai bumped to 0.10.1
+  * devDependencies
+    * @aigne/agent-library bumped to 1.21.1
+    * @aigne/core bumped to 1.35.0
+    * @aigne/test-utils bumped to 0.5.9
+    * @aigne/default-memory bumped to 1.0.1
+
+## [0.10.0](https://github.com/AIGNE-io/aigne-framework/compare/transport-v0.9.6...transport-v0.10.0) (2025-07-15)
+
+
+### Features
+
+* **memory:** support did space memory adapter ([#229](https://github.com/AIGNE-io/aigne-framework/issues/229)) ([6f69b64](https://github.com/AIGNE-io/aigne-framework/commit/6f69b64e98b963db9d6ab5357306b445385eaa68))
+* **model:** support aigne-hub model adapter ([#253](https://github.com/AIGNE-io/aigne-framework/issues/253)) ([4b33f8d](https://github.com/AIGNE-io/aigne-framework/commit/4b33f8d1a819f52357db81d502c56b55eaa0669f))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/openai bumped to 0.10.0
+  * devDependencies
+    * @aigne/agent-library bumped to 1.21.0
+    * @aigne/core bumped to 1.34.0
+    * @aigne/test-utils bumped to 0.5.8
+    * @aigne/default-memory bumped to 1.0.0
+
 ## [0.9.6](https://github.com/AIGNE-io/aigne-framework/compare/transport-v0.9.5...transport-v0.9.6) (2025-07-14)